+++ src/main/java/com/library/librarytask/repository/TitleRepository.java
@@ -9,13 +9,13 @@ import java.util.Optional;
 
 
 @Repository
-public interface TitleRepository extends CrudRepository<Title, Integer> {
+public interface TitleRepository extends CrudRepository<Title, Long> {
 
     @Override
     List<Title> findAll();
 
     @Override
-    Optional<Title> findById(Integer id);
+    Optional<Title> findById(Long id);
 
     @Override
     Title save(final Title title);
